Make sure there is a bare Exifdata object always
It appears that some files don't actually contain any exif data.
Creating an object with such a file would put a type object into
the $!exif attribute, causing later havoc.

This commit ensures that that is at least a bare ExifData object
created when loading a file that doesn't contain any exif data.

An example of such a file is:

   nqp/MoarVM/3rdparty/libuv/docs/src/static/architecture.png

in any Rakudo installation.

Fixes issue #1